requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
click==6.7
configparser==3.5.0
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.6
Unidecode==0.4.20
Werkzeug==0.12.1
